chore(lint): add golangci-lint config and fix existing findings
Add a curated golangci-lint configuration with the default linters
(errcheck, govet, staticcheck, unused, ineffassign, gosimple) plus
correctness and style checks (errorlint, nilerr, bodyclose, gocritic,
revive, copyloopvar, prealloc, misspell, unconvert, gofmt, goimports).

revive's `exported` rule (require godoc on every exported identifier)
is intentionally omitted: it is documentation hygiene best added
incrementally, not enforced retroactively. gocritic's rangeValCopy is
disabled because Spec is by-value by design.

Fix all 19 findings the linter surfaces against the current tree:

- internal/ipc/acktracker.go, pkg/event/event.go: rename `max` local
  to `maxSeq` / `sizeMax` to stop shadowing the Go 1.21+ builtin
- internal/agent/{dispatcher,main}.go, internal/controller/controller.go:
  drop unnecessary uint8(ipc.ProtocolVersion) conversions
  (ProtocolVersion is already `byte`)
- internal/agent/dispatcher.go: stopChild's timeout parameter is
  enforced inside Runner via StopTimeout, not re-applied here. Mark
  it as `_ time.Duration` and document why; remove dead default-set.
- pkg/sandbox/provider_test.go: drop unused `cmd *exec.Cmd` field
- test/integration/shutdown_test.go: drop redundant Go 1.22+
  `h := h` loop-variable capture
- pkg/sandbox/providers/scriptdir/scriptdir.go: split joined path
  argument into separate filepath.Join components
- cmd/cliwrap-agent/main.go: refactor main into main+run helper so
  defer cancel() actually runs (was unreachable due to os.Exit)
- internal/mgmt/server.go, cmd/cliwrap/cmd_{list,status,stop}.go:
  wrap `defer cli.Close()` in a closure that explicitly discards the
  error, satisfying errcheck while preserving intent
- pkg/cliwrap/manager_events.go: replace internal type assertion
  `m.Events().(*eventbus.Bus)` with a typed helper `ensureBus()`
  that returns the concrete type directly

All build / vet / unit / integration / chaos targets remain green.

+++ internal/agent/main.go
@@ -75,7 +75,7 @@ func Run(ctx context.Context, cfg Config) error {
 	conn.Start()
 
 	// Send initial HELLO.
-	helloPayload := ipc.HelloPayload{ProtocolVersion: uint8(ipc.ProtocolVersion), AgentID: cfg.AgentID}
+	helloPayload := ipc.HelloPayload{ProtocolVersion: ipc.ProtocolVersion, AgentID: cfg.AgentID}
 	if err := d.SendControl(ipc.MsgHello, helloPayload, false); err != nil {
 		return fmt.Errorf("agent: send hello: %w", err)
 	}
